Explicitly state `EOTF` in HLG and PQ stanzas
To avoid confusion, this commit modifies the config to explicitly state it's using HLG and PQ's EOTF and the inverse only.

I believe the logic of the whole OETF, EOTF, OOTF concept is that:
- OOTF is basically OETF + EOTF.
- OETF is basically a `view transform` the standard document designed for cameras. 
- The standard document also mentioned that for technical conversions, such as conversion between HLG and PQ encodings, we should use the Inverse EOTF, not the OETF.

Just to add to the weirdness, OCIO's built-in transform for HLG's OETF seems to have a pre-division to the linear RGB values before sending to the curve, I looked up the code, the reason seems to be, to make linear 0.18 map to HLG encoded value of 0.42. For reference, the inverse EOTF HLG curve scaled for SDR peak = 203 nits seems to have 0.18 mapping to about 0.43. So I kind of understand what they are doing there. But still, this whole OETF situation is too weird.

Therefore, this config explicitly use only the EOTF and its inverse.